Platform handling of start and end time stamps

+++ inst/sql/sql_server/CreateCohortTables.sql
@@ -98,7 +98,7 @@
   CREATE TABLE @cohort_database_schema.@cohort_checksum_table(
     cohort_definition_id int NOT NULL,
     checksum varchar(500) NOT NULL,
-    start_time BIGINT,
-    end_time BIGINT
+    start_time FLOAT,
+    end_time FLOAT
   );
 }:{}
